+++ src/main/java/org/polypheny/simpleclient/executor/PolyphenyDbRestExecutor.java
@@ -111,7 +111,7 @@ public class PolyphenyDbRestExecutor implements PolyphenyDbExecutor {
             } catch ( ExecutorException e ) {
                 throw new ExecutorException( "Error while executing query via JDBC", e );
             } finally {
-                commitAndCloseJdbcExecutor( executor );
+                PolyphenyDbJdbcExecutor.commitAndCloseJdbcExecutor( executor );
             }
         }
 
@@ -166,7 +166,7 @@ public class PolyphenyDbRestExecutor implements PolyphenyDbExecutor {
             } catch ( ExecutorException e ) {
                 throw new ExecutorException( "Error while executing query via JDBC", e );
             } finally {
-                commitAndCloseJdbcExecutor( executor );
+                PolyphenyDbJdbcExecutor.commitAndCloseJdbcExecutor( executor );
             }
         }
     }
@@ -227,7 +227,7 @@ public class PolyphenyDbRestExecutor implements PolyphenyDbExecutor {
         } catch ( ExecutorException e ) {
             throw new ExecutorException( "Error while executing query via JDBC", e );
         } finally {
-            commitAndCloseJdbcExecutor( executor );
+            PolyphenyDbJdbcExecutor.commitAndCloseJdbcExecutor( executor );
         }
     }
 
@@ -242,7 +242,7 @@ public class PolyphenyDbRestExecutor implements PolyphenyDbExecutor {
         } catch ( ExecutorException e ) {
             throw new ExecutorException( "Error while executing query via JDBC", e );
         } finally {
-            commitAndCloseJdbcExecutor( executor );
+            PolyphenyDbJdbcExecutor.commitAndCloseJdbcExecutor( executor );
         }
     }
 
@@ -258,30 +258,9 @@ public class PolyphenyDbRestExecutor implements PolyphenyDbExecutor {
             log.error( "Exception while setting config \"" + key + "\"!", e );
         } finally {
             try {
-                commitAndCloseJdbcExecutor( executor );
+                PolyphenyDbJdbcExecutor.commitAndCloseJdbcExecutor( executor );
             } catch ( ExecutorException e ) {
                 log.error( "Exception while closing JDBC executor", e );
-            }
-        }
-    }
-
-
-    public static void commitAndCloseJdbcExecutor( JdbcExecutor executor ) throws ExecutorException {
-        if ( executor != null ) {
-            try {
-                executor.executeCommit();
-            } catch ( ExecutorException e ) {
-                try {
-                    executor.executeRollback();
-                } catch ( ExecutorException ex ) {
-                    log.error( "Error while rollback connection", e );
-                }
-            } finally {
-                try {
-                    executor.closeConnection();
-                } catch ( ExecutorException e ) {
-                    log.error( "Error while closing connection", e );
-                }
             }
         }
     }
